+++ test/test_type_definition.ts
@@ -33,6 +33,7 @@ var None = OptionT.None;
 (function(){
     var option: OptionT.Some<number> = new Some(1);
     var isSome: boolean = option.isSome;
+    var isNone: boolean = option.isNone;
     var unwrap: number = option.unwrap();
     var unwrapOr: number = option.unwrapOr(10);
     var map: OptionT.Option<string> = option.map((val: number): string => String(val));
@@ -54,6 +55,7 @@ var None = OptionT.None;
 (function(){
     var option: OptionT.None<number> = new None<number>();
     var isSome: boolean = option.isSome;
+    var isNone: boolean = option.isNone;
     var unwrap: number = option.unwrap();
     var unwrapOr: number = option.unwrapOr(10);
     var map: OptionT.Option<string> = option.map((val: number): string => String(val));